refactor(ToolbarComponent): to resolve layout issue with sidenav
