+++ src/app/app.component.scss
@@ -17,3 +17,13 @@
   display: flex;
   flex-direction: column;
 }
+
+mat-sidenav-container {
+  flex: 1 1 auto;
+  width: 100%;
+  height: calc(100vh - 64px);
+}
+
+mat-sidenav {
+  width: 200px;
+}
