+++ src/app/app.component.ts
@@ -1,8 +1,37 @@
 import { Component } from "@angular/core";
 
+interface Menu {
+  title: string;
+  link: string;
+  options: { exact: boolean };
+}
+
 @Component({
   selector: "app-root",
   templateUrl: "./app.component.html",
   styleUrls: ["./app.component.scss"]
 })
-export class AppComponent {}
+export class AppComponent {
+  menus: Menu[] = [
+    {
+      title: "Home",
+      link: "/",
+      options: { exact: true }
+    },
+    {
+      title: "Operators",
+      link: "/operators",
+      options: { exact: false }
+    },
+    {
+      title: "Companies",
+      link: "/companies",
+      options: { exact: false }
+    },
+    {
+      title: "Team",
+      link: "/team",
+      options: { exact: false }
+    }
+  ];
+}
