+++ src/app/companies/companies.component.html
@@ -1,3 +1,1 @@
-<p>
-  companies works!
-</p>
+
