+++ src/app/toolbar/toolbar.component.scss
@@ -6,13 +6,3 @@ img {
   padding: 0 16px;
   font-weight: 600;
 }
-
-mat-sidenav-container {
-  flex: 1 1 auto;
-  width: 100%;
-  height: calc(100vh - 64px);
-}
-
-mat-sidenav {
-  width: 200px;
-}
