+++ src/app/toolbar/toolbar.component.ts
@@ -1,10 +1,4 @@
-import { Component, OnInit } from "@angular/core";
-
-interface Menu {
-  title: string;
-  link: string;
-  options: { exact: boolean };
-}
+import { Component, OnInit, Output, EventEmitter } from "@angular/core";
 
 @Component({
   selector: "app-toolbar",
@@ -12,28 +6,10 @@ interface Menu {
   styleUrls: ["./toolbar.component.scss"]
 })
 export class ToolbarComponent implements OnInit {
-  menus: Menu[] = [
-    {
-      title: "Home",
-      link: "/",
-      options: { exact: true }
-    },
-    {
-      title: "Operators",
-      link: "/operators",
-      options: { exact: false }
-    },
-    {
-      title: "Companies",
-      link: "/companies",
-      options: { exact: false }
-    },
-    {
-      title: "Team",
-      link: "/team",
-      options: { exact: false }
-    }
-  ];
+  @Output() navToggle = new EventEmitter<boolean>();
+  navOpen() {
+    this.navToggle.emit(true);
+  }
 
   constructor() {}
